assets/hpc/hf/lora/requirements.txt
accelerate==1.6.0 ; python_version == "3.12"
aiohappyeyeballs==2.6.1 ; python_version == "3.12"
aiohttp==3.11.14 ; python_version == "3.12"
aiosignal==1.3.2 ; python_version == "3.12"
airportsdata==20250523 ; python_version == "3.12"
alembic==1.15.2 ; python_version == "3.12"
annotated-types==0.7.0 ; python_version == "3.12"
attrs==25.1.0 ; python_version == "3.12"
certifi==2025.1.31 ; python_version == "3.12"
charset-normalizer==3.4.1 ; python_version == "3.12"
click==8.1.8 ; python_version == "3.12"
cloudpickle==3.1.1 ; python_version == "3.12"
colorama==0.4.6 ; (platform_system == "Windows" or sys_platform == "win32") and python_version == "3.12"
colorlog==6.9.0 ; python_version == "3.12"
datasets==3.5.0 ; python_version == "3.12"
dill==0.3.8 ; python_version == "3.12"
diskcache==5.6.3 ; python_version == "3.12"
docker-pycreds==0.4.0 ; python_version == "3.12"
filelock==3.17.0 ; python_version == "3.12"
frozenlist==1.5.0 ; python_version == "3.12"
fsspec==2024.12.0 ; python_version == "3.12"
genson==1.3.0 ; python_version == "3.12"
gitdb==4.0.12 ; python_version == "3.12"
gitpython==3.1.44 ; python_version == "3.12"
greenlet==3.2.1 ; python_version == "3.12" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32")
huggingface-hub==0.30.2 ; python_version == "3.12"
idna==3.10 ; python_version == "3.12"
interegular==0.3.3 ; python_version == "3.12"
iso3166==2.1.1 ; python_version == "3.12"
jinja2==3.1.5 ; python_version == "3.12"
joblib==1.5.1 ; python_version == "3.12"
jsonschema-specifications==2024.10.1 ; python_version == "3.12"
jsonschema==4.23.0 ; python_version == "3.12"
lark==1.2.2 ; python_version == "3.12"
mako==1.3.10 ; python_version == "3.12"
markdown-it-py==3.0.0 ; python_version == "3.12"
markupsafe==3.0.2 ; python_version == "3.12"
mdurl==0.1.2 ; python_version == "3.12"
mpmath==1.3.0 ; python_version == "3.12"
multidict==6.2.0 ; python_version == "3.12"
multiprocess==0.70.16 ; python_version == "3.12"
nest-asyncio==1.6.0 ; python_version == "3.12"
networkx==3.4.2 ; python_version == "3.12"
numpy==2.2.5 ; python_version == "3.12"
nvidia-cublas-cu12==12.4.5.8 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version == "3.12"
nvidia-cuda-cupti-cu12==12.4.127 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version == "3.12"
nvidia-cuda-nvrtc-cu12==12.4.127 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version == "3.12"
nvidia-cuda-runtime-cu12==12.4.127 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version == "3.12"
nvidia-cudnn-cu12==9.1.0.70 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version == "3.12"
nvidia-cufft-cu12==11.2.1.3 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version == "3.12"
nvidia-curand-cu12==10.3.5.147 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version == "3.12"
nvidia-cusolver-cu12==11.6.1.9 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version == "3.12"
nvidia-cusparse-cu12==12.3.1.170 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version == "3.12"
nvidia-cusparselt-cu12==0.6.2 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version == "3.12"
nvidia-nccl-cu12==2.21.5 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version == "3.12"
nvidia-nvjitlink-cu12==12.4.127 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version == "3.12"
nvidia-nvtx-cu12==12.4.127 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version == "3.12"
optuna==4.3.0 ; python_version == "3.12"
outlines-core==0.1.26 ; python_version == "3.12"
outlines==0.2.3 ; python_version == "3.12"
packaging==24.2 ; python_version == "3.12"
pandas==2.2.3 ; python_version == "3.12"
peft==0.15.1 ; python_version == "3.12"
platformdirs==4.3.6 ; python_version == "3.12"
propcache==0.3.1 ; python_version == "3.12"
protobuf==5.29.3 ; python_version == "3.12" and sys_platform != "linux"
psutil==7.0.0 ; python_version == "3.12"
pyarrow==19.0.1 ; python_version == "3.12"
pydantic-core==2.33.2 ; python_version == "3.12"
pydantic==2.11.4 ; python_version == "3.12"
pygments==2.19.1 ; python_version == "3.12"
python-dateutil==2.9.0.post0 ; python_version == "3.12"
python-decouple==3.8 ; python_version == "3.12"
pytz==2025.1 ; python_version == "3.12"
pyyaml==6.0.2 ; python_version == "3.12"
referencing==0.36.2 ; python_version == "3.12"
regex==2024.11.6 ; python_version == "3.12"
requests==2.32.3 ; python_version == "3.12"
rich==14.0.0 ; python_version == "3.12"
rpds-py==0.22.3 ; python_version == "3.12"
safetensors==0.5.3 ; python_version == "3.12"
scikit-learn==1.6.1 ; python_version == "3.12"
scipy==1.15.3 ; python_version == "3.12"
sentencepiece==0.2.0 ; python_version == "3.12"
sentry-sdk==2.26.1 ; python_version == "3.12"
setproctitle==1.3.5 ; python_version == "3.12"
setuptools==75.8.0 ; python_version == "3.12"
six==1.17.0 ; python_version == "3.12"
smmap==5.0.2 ; python_version == "3.12"
sqlalchemy==2.0.40 ; python_version == "3.12"
sympy==1.13.1 ; python_version == "3.12"
threadpoolctl==3.6.0 ; python_version == "3.12"
tokenizers==0.21.1 ; python_version == "3.12"
torch==2.6.0 ; python_version == "3.12"
tqdm==4.67.1 ; python_version == "3.12"
transformers==4.51.3 ; python_version == "3.12"
triton==3.2.0 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version == "3.12"
trl==0.16.1 ; python_version == "3.12"
typing-extensions==4.12.2 ; python_version == "3.12"
typing-inspection==0.4.0 ; python_version == "3.12"
tzdata==2025.2 ; python_version == "3.12"
urllib3==2.3.0 ; python_version == "3.12"
wandb==0.19.9 ; python_version == "3.12"
xxhash==3.5.0 ; python_version == "3.12"
yarl==1.18.3 ; python_version == "3.12"
